% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.r
\name{tidy_html}
\alias{tidy_html}
\title{Tidy HTML/XML/XHTML Documents}
\usage{
tidy_html(content, options = list(TidyXhtmlOut = TRUE))
}
\arguments{
\item{content}{atomic character or raw vector of content to tidy}

\item{options}{named list of options}
}
\value{
atomic character vector of tidy content
}
\description{
Tidy HTML/XML/XHTML Documents
}

