% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.r
\docType{data}
\name{tidy_options}
\alias{tidy_options}
\title{HTML, XHTML & XML Options for tidy_html}
\format{A data frame with 55 rows and 3 variables}
\usage{
tidy_options
}
\description{
This dataset contains the options (and their default settings) for
tidy_html. They are passed in a named-list to tidy_html

\itemize{
  \item \code{Option}: Option name
  \item \code{Type}: Option value type
  \item \code{Default}: Is it the default for tidy_html?
}
}
\note{
Last updated 2016-09-09.
}
\references{
The \href{http://api.html-tidy.org/tidy/quickref_5.1.25.html}{
  HTML Tidy Options Quick Reference}
}
\keyword{datasets}

