% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltidy-package.r
\docType{package}
\name{htmltidy}
\alias{htmltidy}
\alias{htmltidy-package}
\title{Clean Up Gnarly HTML/XML}
\description{
HTML and XML documents can be beautiful and pristine. They can also be
wretched, evil, malformed hellspawn. Now, you can tidy up that HTML and XML before
processing it with your favorite angle-bracket parsing tools.
}
\author{
Bob Rudis (bob@rud.is)
}

