% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tidy_html}
\alias{tidy_html}
\title{Tidy HTML/XML}
\usage{
tidy_html(source)
}
\arguments{
\item{source}{length 1 character vetor containing the HTML/XML source to process}
}
\description{
Tidy HTML/XML
}

